fix of comments
