+++ components/bl-simpleModal-component/src/index.js
@@ -1,33 +1,24 @@
-import { useState, useEffect } from 'react';
-import { Button, Container, Title } from './subcomponents';
-import modalTypes from './modal-types';
+import { useState } from 'react';
+import { Modal } from './modal';
+
+const { cn } = BackendlessUI.CSSUtils;
 
 export default function SimpleModal({ component, eventHandlers }) {
   const {
     display,
-    classList,
-    title,
-    content,
-    type,
-    placeholder,
-    closeButtonLabel,
-    submitButtonLabel,
     closingDuration,
   } = component;
-  const { onClose, onSubmit, onInputValueChange } = eventHandlers;
 
   const [isOpen, setIsOpen] = useState(false);
   const [isClosing, setIsClosing] = useState(false);
   const [inputValue, setInputValue] = useState('');
-
-  const { rootClasses, modalClasses } = useClasses(classList, isClosing);
 
   component.closeModal = () => {
     setIsClosing(true);
 
     setTimeout(() => {
       setIsOpen(false);
-      setIsClosing(false)
+      setIsClosing(false);
       setInputValue('');
     }, closingDuration);
   };
@@ -36,53 +27,18 @@ export default function SimpleModal({ component, eventHandlers }) {
     setIsOpen(true);
   };
 
-  useEffect(() => {
-    onInputValueChange({ inputValue });
-  }, [inputValue]);
-
-  useEffect(() => {
-    document.body.classList.toggle('active-modal', isOpen);
-  }, [isOpen]);
-
   if (!display || !isOpen) {
     return null;
   }
 
   return (
-    <div className={ rootClasses }>
-      <div className={ modalClasses } style={ { animationDuration: `${ closingDuration }ms` } }>
-        <div onClick={ onClose } className="overlay"></div>
-        <div className="simple-modal__content">
-          <Title content={ title }/>
-          { (type === modalTypes.prompt || content) && (
-            <Container
-              content={ content }
-              type={ type }
-              inputValue={ inputValue }
-              setInputValue={ setInputValue }
-              placeholder={ placeholder }
-            />
-          ) }
-
-          <Button
-            type={ type }
-            onClose={ onClose }
-            onSubmit={ onSubmit }
-            inputValue={ inputValue }
-            submitButtonLabel={ submitButtonLabel }
-            closeButtonLabel={ closeButtonLabel }
-          />
-        </div>
-      </div>
-    </div>
+    <Modal
+      component={ component }
+      eventHandlers={ eventHandlers }
+      inputValue={ inputValue }
+      isOpen={ isOpen }
+      setInputValue={ setInputValue }
+      isClosing={ isClosing }
+    />
   );
 }
-
-const useClasses = (classList, isClosing) => {
-  const rootClasses = ['bl-customComponent-simple-modal ', ...classList];
-  const modalClasses = ['simple-modal'];
-
-  modalClasses.push(isClosing ? 'close-modal' : 'open-modal');
-
-  return { rootClasses: rootClasses.join(' '), modalClasses: modalClasses.join(' ') };
-};
