+++ components/bl-simpleModal-component/styles/index.less
@@ -16,7 +16,7 @@ body.active-modal {
 }
 
 .bl-customComponent-simple-modal {
-  z-index: 1;
+  z-index: 1000;
 
   .overlay {
     position: fixed;
